==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
# 1. Load data

import pandas as pd

rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')
rides.head()

# 2. Tranformations

## Round previous hour
rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

## Rides by hours
agg_rides = rides.groupby(['pickup_hour','pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0:'rides'},inplace=True)
agg_rides

agg_rides.loc[agg_rides['pickup_location_id'] == 2,['pickup_hour','rides']]

agg_rides_i =  agg_rides.loc[agg_rides['pickup_location_id'] == 2,['pickup_hour','rides']]

agg_rides_i.set_index('pickup_hour',inplace=True)

pd.DatetimeIndex(agg_rides_i.index)

full_range = pd.date_range(agg_rides['pickup_hour'].min(),agg_rides['pickup_hour'].max(),freq='H')

agg_rides_i['pickup_location_id'] = 2

agg_rides_i

## Define hours with 0 Rides
from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(agg_rides['pickup_hour'].min(),agg_rides['pickup_hour'].max(),freq='H')
    output = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        # Get only dates with rides
        agg_rides_i =  agg_rides.loc[agg_rides['pickup_location_id'] == location_id,['pickup_hour','rides']]
        
        # Add missing dates with 0
        agg_rides_i.set_index('pickup_hour',inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range,fill_value=0)
        
        # Add location ids to dates and concatenate
        agg_rides_i['pickup_location_id'] = location_id
        output = pd.concat([output,agg_rides_i])
    
    output = output.reset_index().rename(columns={'index':'pickup_hour'})
    
    return output
        

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

# 3. Plot time series data
* Se usa para detectar errores en el proceso de transformación
* Plotly me permite interactuar con las graficas

from typing import Optional, List
import plotly.express as px 

def plot_rides(rides: pd.DataFrame, locations:Optional[List[int]] = None):
    # Plot time-series data
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides
    
    fig = px.line(
        rides_to_plot,
        x='pickup_hour',
        y='rides',
        color='pickup_location_id',
        template='none',
    )
    
    fig.show()

plot_rides(agg_rides_all_slots,locations=[4,17])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
